Refactor redirect storage for magic links
https://github.com/communitiesuk/funding-service/pull/73#discussion_r2042295417

Let's store redirect/next URLs in the session for more
tamper-resistance, since the session is set and signed by the server -
so let's able to be interfered with by users.
