+++ tests/integration/common/auth/test_auth.py
@@ -1,5 +1,6 @@
 import datetime
 
+import pytest
 from bs4 import BeautifulSoup
 from flask import url_for
 from flask_login import current_user
@@ -45,16 +46,26 @@ class TestSignInView:
             == "http://funding.communities.gov.localhost:8080/request-a-link-to-sign-in"
         )
 
-    def test_post_valid_email_with_redirect(self, client, mock_notification_service_calls, db_session):
+    @pytest.mark.parametrize(
+        "next_, safe_next",
+        (
+            ("/blah/blah", "/blah/blah"),
+            ("https://bad.place/blah", "/"),  # Single test case; see TestSanitiseRedirectURL for more exhaustion
+        ),
+    )
+    def test_post_valid_email_with_redirect(
+        self, client, mock_notification_service_calls, db_session, next_, safe_next
+    ):
+        with client.session_transaction() as session:
+            session["next"] = next_
+
         response = client.post(
-            url_for("auth.request_a_link_to_sign_in", next="/blah/blah"),
+            url_for("auth.request_a_link_to_sign_in"),
             data={"email_address": "test@communities.gov.uk"},
             follow_redirects=True,
         )
         assert response.status_code == 200
-        assert (
-            db_session.scalar(select(MagicLink).order_by(MagicLink.created_at.desc())).redirect_to_path == "/blah/blah"
-        )
+        assert db_session.scalar(select(MagicLink).order_by(MagicLink.created_at.desc())).redirect_to_path == safe_next
 
 
 class TestCheckEmailPage:
@@ -101,9 +112,16 @@ class TestClaimMagicLinkView:
         assert response.status_code == 302
         assert response.location == url_for("auth.request_a_link_to_sign_in")
 
-    def test_post_claims_link_and_redirects(self, client, factories):
+    @pytest.mark.parametrize(
+        "redirect_to, safe_redirect_to",
+        (
+            ("/blah/blah", "/blah/blah"),
+            ("https://bad.place/blah", "/"),  # Single test case; see TestSanitiseRedirectURL for more exhaustion
+        ),
+    )
+    def test_post_claims_link_and_redirects(self, client, factories, redirect_to, safe_redirect_to):
         magic_link = factories.magic_link.create(
-            user__email="test@communities.gov.uk", redirect_to_path="/my-redirect", claimed_at_utc=None
+            user__email="test@communities.gov.uk", redirect_to_path=redirect_to, claimed_at_utc=None
         )
 
         assert current_user.is_authenticated is False
@@ -115,7 +133,7 @@ class TestClaimMagicLinkView:
         )
 
         assert response.status_code == 302
-        assert response.location == "/my-redirect"
+        assert response.location == safe_redirect_to
         assert magic_link.claimed_at_utc is not None
         assert current_user.is_authenticated is True
         assert magic_link.user.id == current_user.id
